Add GNOME as a configurable option in nixos configurations
This commit introduces the ability to specify GNOME as a configurable option in 'flake.nix' and 'default.nix'. This allows users to set the desktop environment to GNOME in 'flake.nix', which will then be automatically used in 'default.nix'. This increases code flexibility, enabling easy swapping between different desktop environments.