Add 'config' to function arguments in `kde.nix`
The updated version of `kde.nix` now includes 'config' as an argument to the initial function declared within it. This was necessary because the `programs.plasma` settings were moved from the home manager file to this user environment, making it essential to use the 'config' reference. The `plasma-manager` import was also transferred to this file to enable easier module management.