Add neovim configuration for LazyVim
This commit adds the configuration for the LazyVim Neovim setup. It includes the various parts of the setup including a .gitignore file, a config directory with auto commands, keymaps, options, and a lazy.lua file. It also includes a LICENSE file, README file, and a stylua.toml file for Lua code formatting. The configuration was added to the home-manager and system configurations. The 'nixfmt' was also removed from systemPackages as it was unnecessary in this context.